+++ src/runtime/components/ChatPromptSubmit.vue
@@ -58,8 +58,8 @@ export interface ChatPromptSubmitProps extends Omit<ButtonProps, 'icon' | 'color
 }
 
 export interface ChatPromptSubmitEmits {
-  stop: []
-  reload: []
+  stop: [event: MouseEvent]
+  reload: [event: MouseEvent]
 }
 </script>
 
@@ -100,25 +100,25 @@ const statusButtonProps = computed(() => ({
   submitted: {
     icon: props.submittedIcon || icons.stop,
     color: props.submittedColor,
-    onClick() {
-      emits('stop')
+    onClick(e) {
+      emits('stop', e)
     }
   },
   streaming: {
     icon: props.streamingIcon || icons.stop,
     color: props.streamingColor,
-    onClick() {
-      emits('stop')
+    onClick(e) {
+      emits('stop', e)
     }
   },
   error: {
     icon: props.errorIcon || icons.reload,
     color: props.errorColor,
-    onClick() {
-      emits('reload')
+    onClick(e) {
+      emits('reload', e)
     }
   }
-})[props.status])
+} satisfies { [key: string]: ButtonProps })[props.status])
 
 // eslint-disable-next-line vue/no-dupe-keys
 const b24ui = computed(() => tv({ extend: tv(theme), ...(appConfig.b24ui?.chatPromptSubmit || {}) })())