+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * docs(ComponentCode): add missing cast imports
 * docs(LLMs): fix typo
 * fix(module): put back `#build/ui.css` alias
+* fix(ChatPromptSubmit): proxy event to `stop` and `reload` emits
 
 ## [2.1.6](https://github.com/bitrix24/b24ui/compare/v2.1.5...v2.1.6) (2025-11-20)